finish Checkout and stripe Payment Integration

+++ Models/CinemaM/Promotion.cs
@@ -6,7 +6,7 @@
         public string Code { get; set; } = string.Empty;
         public string Description { get; set; } = string.Empty;
 
-        public int DiscountPercentage { get; set; } 
+        public decimal DiscountPercentage { get; set; } 
         public DateTime ValidFrom { get; set; }
         public DateTime ValidTo { get; set; }
